+++ src/x86/isr.h
@@ -1,0 +1,19 @@
+#pragma once
+
+__attribute__ ((interrupt))
+void genericISR(void* frame);
+
+__attribute__ ((interrupt))
+void divisionByZero(void* frame);
+
+__attribute__ ((interrupt))
+void gpFaultHandler(void* frame, unsigned int error);
+
+__attribute__ ((interrupt))
+void pageFaultHandler(void* frame, unsigned int error);
+
+__attribute__ ((interrupt))
+void doubleFaultHandler(void* frame, unsigned int error);
+
+__attribute__ ((interrupt))
+void syscallHandler(void* frame);
